tongqu-agent-backend/requirements.txt
# 核心
python-dotenv>=1.0.0
requests>=2.31.0

# Gemini：仅绘本配图（文生图）；OpenAI 兼容中转需 openai
google-genai>=1.0.0
openai>=1.0.0

# HTTP 服务
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# 阿里云百炼（故事）+ Green + NLS
dashscope>=1.25.6
alibabacloud-tea-openapi>=0.3.8
alibabacloud-tea-util>=0.3.11
alibabacloud-green20220302>=2.0.0
aliyun-python-sdk-core>=2.14.0
